bot/requirements.txt
asyncpg==0.29.0
discord.py==2.3.2
